Add builder creation prior to building images on buildkit
The `buildkit` tests are sometimes flaky due to the fact that we are
trying to create 2 images in parallel. When we try to do this `buildkit`
will try to create a builder, if it doesn't exist, in parallel which
causes a failure.

This commit ensures that the builder is created before we try to run the
parallel test to prevent the above-described issue