fix: trim 'v' prefix from version string
When kbld is used as a Go module dependency, the version string is taken from
the Go module version, which currently includes a 'v' prefix (e.g., 'v0.46.0').
The validation logic for the kbld config's minimumRequiredVersion field does not
accept this 'v' prefix, leading to validation errors like the following:

  kbld: Error: Validating config/ (kbld.k14s.io/v1alpha1) cluster:
    Validating minimum version:
      Must not have prefix 'v' (e.g. '0.8.0')

This change trims the 'v' prefix from the version string returned from
dependecies.

Signed-off-by: German Lashevich <german.lashevich@gmail.com>